+++ CHANGES.rst
@@ -17,6 +17,11 @@ Unreleased
   ``history_list_display`` by default, and made the latter into an actual field (gh-1128)
 - ``ModelDelta`` and ``ModelChange`` (in ``simple_history.models``) are now immutable
   dataclasses; their signatures remain unchanged (gh-1128)
+- ``ModelDelta``'s ``changes`` and ``changed_fields`` are now sorted alphabetically by
+  field name. Also, if ``ModelChange`` is for an M2M field, its ``old`` and ``new``
+  lists are sorted by the related object. This should help prevent flaky tests. (gh-1128)
+- ``diff_against()`` has a new keyword argument, ``foreign_keys_are_objs``;
+  see usage in the docs under "History Diffing" (gh-1128)
 
 3.5.0 (2024-02-19)
 ------------------